refactor: 개별 쿠폰 선택 로직 Hook으로 분리
