+++ src/components/CartPage/CouponSelect/SpecificCouponSelect.tsx
@@ -1,12 +1,5 @@
-import { useRecoilState, useRecoilValue } from 'recoil';
 import { Coupon, CouponSelect } from './SpecificCouponSelect.styles';
-import {
-  couponsSelector,
-  selectedCouponsState,
-  specificCouponSelector,
-} from '../../../atoms/coupons';
-import { useEffect, useState } from 'react';
-import { isSelectedCartId } from '../../../atoms/cart';
+import useCouponSelect from '../../../hooks/useCouponSelect';
 
 interface SpecificCouponSelectProps {
   cartId: number;
@@ -17,39 +10,9 @@ const SpecificCouponSelect = ({
   cartId,
   productId,
 }: SpecificCouponSelectProps) => {
-  const { allCoupons } = useRecoilValue(couponsSelector);
-  const specificCoupons = useRecoilValue(specificCouponSelector(productId));
-  const [selectedCoupons, setSelectedCoupons] =
-    useRecoilState(selectedCouponsState);
-  const [targetCouponId, setTargetCouponId] = useState<number>();
-  const isCheckedCartItem = useRecoilValue(isSelectedCartId(cartId));
+  const { selectedCoupons, specificCoupons, onCouponSelected } =
+    useCouponSelect(cartId, productId);
 
-  useEffect(() => {
-    if (!isCheckedCartItem && targetCouponId) {
-      setSelectedCoupons(
-        selectedCoupons.filter((couponId) => couponId !== targetCouponId)
-      );
-    }
-  }, [isCheckedCartItem]);
-
-  const onCouponSelected = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    if (e.target.value === 'DEFAULT' || !isCheckedCartItem) {
-      e.target.selectedIndex = 0;
-      setSelectedCoupons(
-        selectedCoupons.filter((couponId) => couponId !== targetCouponId)
-      );
-      return;
-    }
-
-    const newSelectedCoupons = selectedCoupons.filter(
-      (couponId) => !allCoupons.some((coupon) => coupon.id === couponId)
-    );
-    const newSelectedCouponId = +e.target.value;
-    newSelectedCoupons.push(newSelectedCouponId);
-
-    setTargetCouponId(newSelectedCouponId);
-    setSelectedCoupons([...new Set(newSelectedCoupons)]);
-  };
   return (
     <CouponSelect onChange={onCouponSelected}>
       <Coupon defaultChecked value={'DEFAULT'}>
